+++ README.md
@@ -1,2 +1,5 @@
-# rbtree
-A red-black tree implementation for learning purposes
+# collections-ext
+This repo is intended to function as a catch-all bucket for datastructures that I am working with for side projects and learning purposes.
+
+## Warning
+Please nobody use this. This is entirely an experiment to support insane restrictions I've imposed on myself to build a computer from first principles.
