+++ examples/tree/basic.rs
@@ -1,4 +1,4 @@
-use rbtree::RedBlackTree;
+use collections_ext::RedBlackTree;
 
 fn main() {
     let populated_tree = (0..1024).fold(RedBlackTree::default(), |tree, x| tree.insert(x));
