+++ src/tree/mod.rs
@@ -1,0 +1,3 @@
+//! Provides a catch-all for tree implementations not currently implemented in stdlib.
+
+pub mod rbtree;
